Move component Filter to Homes
