+++ src/Header/index.js
@@ -9,7 +9,7 @@ const Header = styled.header`
   position: fixed;
   box-sizing: border-box;
   box-shadow: 0px 0.03rem 0px rgba(72, 72, 72, 0.3);
-  z-index: 2;
+  z-index: 3;
 `;
 
 const Wrapper = styled.div`
